# math.big: improve the performance of left_shift_digits_in_place and right_shift_digits_in_place